+++ chrome-remote-cache/src/CacheStat.ts
@@ -5,9 +5,23 @@ import { formatSize, splitUrl } from './CacheUtils'
  * a simple class to eveluate data transfert per domains
  */
 export class CacheStat {
-    public query = 0;
-    public transfert = 0;
-    public perDom: { [key: string]: number } = {};
+    #query = 0;
+    #transfert = 0;
+    #perDom: { [key: string]: number } = {};
+
+    /**
+     * get number of query
+     */
+    public get query(): number {
+        return this.#query;
+    }
+
+    /**
+     * get total bandwidh
+     */
+     public get transfert(): number {
+        return this.#transfert;
+    }
 
     /**
      * increate data usage from a particular url.
@@ -15,18 +29,18 @@ export class CacheStat {
      * @param meta 
      * @param length 
      */
-    add(url: string, meta: string, length?: number): void {
+    public add(url: string, meta: string, length?: number): void {
         if (url.startsWith('data:'))
             return;
-        this.query++;
+        this.#query++;
         let size = meta.length
         if (length)
             size += length;
-        this.transfert += size;
+        this.#transfert += size;
 
         const [dom] = splitUrl(url);
-        let old = this.perDom[dom] || 0;
-        this.perDom[dom] = old + size;
+        let old = this.#perDom[dom] || 0;
+        this.#perDom[dom] = old + size;
     }
 
     /**
@@ -34,10 +48,10 @@ export class CacheStat {
      * 
      * @param full will spit statistique per domains
      */
-    toString(full?: boolean): string {
-        let out = `${pc.green(formatSize(this.transfert))} in ${pc.green(this.query.toString().padStart(3, ' '))} Query`;
+    public toString(full?: boolean): string {
+        let out = `${pc.green(formatSize(this.#transfert))} in ${pc.green(this.#query.toString().padStart(3, ' '))} Query`;
         if (full) {
-            const asArray = Object.entries(this.perDom);
+            const asArray = Object.entries(this.#perDom);
             asArray.sort((a, b) => b[1] - a[1])
             for (const [dom, size] of asArray) {
                 // if (size > 50000)