+++ chrome-remote-cache/CHANGELOG.md
@@ -1,14 +1,22 @@
 # changelog
 
-## v 0.0.3 (2022-07-28)
+## v0.0.4 (2022-07-28)
+* ChromeRemoteCache.getStats() return CacheStat objects
+* improve CacheStat
+* handle empty body
+* add ChromeRemoteCache.cacheIgnore
+* change debug colors
+* add ChromeRemoteCache.efficency
+
+## v0.0.3 (2022-07-28)
 * do not permit duplicate entry in urlSet
 * add del method in urlSet
 * add extrac counter in urlSet
 * enable ava + add tests for urlSet
 
-## v 0.0.2 (2022-07-28)
+## v0.0.2 (2022-07-28)
 * add methods to ChromeRemoteCache
 * add dropQueryParam, formatSize exports
 
-## v 0.0.1 (2022-07-28)
+## v0.0.1 (2022-07-28)
 * initial release